Thmeyer

+++ sources/executing/render_3d.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   render_3d.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: thmeyer < thmeyer@student.42lyon.fr >      +#+  +:+       +#+        */
+/*   By: thmeyer <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/12 14:41:02 by thmeyer           #+#    #+#             */
-/*   Updated: 2023/06/13 14:20:04 by thmeyer          ###   ########.fr       */
+/*   Updated: 2023/06/14 00:00:09 by thmeyer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,8 +40,8 @@ static void	draw_wall(t_data *back, int x, int y, int max)
 
 void	generate_3d(t_cub *cub)
 {
-	int		wall_height;
-	int		margin;
+	double	wall_height;
+	double	margin;
 	int		x;
 
 	if (cub->imgs->back.img)
@@ -53,7 +53,8 @@ void	generate_3d(t_cub *cub)
 	while (cub->p->ray[++x])
 	{
 		cub->p->ray[x]->dist = fix_fisheye(cub, cub->p->ray[x]->dist, x);
-		wall_height = WIN_HEIGHT / cub->p->ray[x]->dist * 40;
+		wall_height = WIN_HEIGHT / (int)cub->p->ray[x]->dist * 40;
+		// wall_height = WIN_HEIGHT / (int)cub->p->ray[x]->dist;
 		margin = (WIN_HEIGHT - wall_height) / 2;
 		if (margin > 0 && margin < WIN_HEIGHT)
 			draw_wall(&cub->imgs->back, x, margin, margin + wall_height - 1);
